+++ src/main/java/com/task/fooddelivery/service/PhenomenonFeeService.java
@@ -1,11 +1,10 @@
 package com.task.fooddelivery.service;
 
-import com.task.fooddelivery.dto.AirTemperatureFeeDto;
 import com.task.fooddelivery.dto.PhenomenonFeeDto;
-import com.task.fooddelivery.entity.AirTemperatureFee;
 import com.task.fooddelivery.entity.DeliveryMethod;
 import com.task.fooddelivery.entity.PhenomenonFee;
 import com.task.fooddelivery.exception.DeliveryForbiddenException;
+import com.task.fooddelivery.exception.NotFoundException;
 import com.task.fooddelivery.repository.PhenomenonFeeRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -27,7 +26,7 @@ public class PhenomenonFeeService {
         List<BigDecimal> fees = new ArrayList<>(List.of(new BigDecimal(0)));
         if (phenomenon == null) return fees.get(0);
         for (PhenomenonFee rule : method.getPhenomenonFees()) {
-            if (rule.getPhenomenonName().toLowerCase(Locale.ROOT).contains(phenomenon)) {
+            if (phenomenon.toLowerCase(Locale.ROOT).contains(rule.getPhenomenonName().toLowerCase(Locale.ROOT))) {
                 if (rule.isDeliveryForbidden()) throw new DeliveryForbiddenException(phenomenon);
                 fees.add(rule.getFee());
             }
@@ -49,6 +48,7 @@ public class PhenomenonFeeService {
     public void deleteFees(String phenomenonName, String methodName) {
         DeliveryMethod method = methodService.getDeliveryMethodEntity(methodName);
         List<PhenomenonFee> fees = feeRepository.findAllByPhenomenonNameAndDeliveryMethod(phenomenonName, method);
+        if (fees.isEmpty()) throw new NotFoundException("No suitable phenomenon fee rules found.");
         feeRepository.deleteAll(fees);
     }
 }
